+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [3.21.4](https://github.com/witchesofthehill/manabrew/compare/v3.21.3...v3.21.4) (2026-08-25)
+
+## [3.21.3](https://github.com/witchesofthehill/manabrew/compare/v3.21.2...v3.21.3) (2026-08-25)
+
+## [3.21.2](https://github.com/witchesofthehill/manabrew/compare/v3.21.1...v3.21.2) (2026-08-24)
+
+### Fixes
+
+* **node:** report engine stall time under g1 ([#757](https://github.com/witchesofthehill/manabrew/issues/757)) ([f16022f](https://github.com/witchesofthehill/manabrew/commit/f16022febfe14d4e1a31f5f094b71227bf3958a8))
+
+## [3.21.1](https://github.com/witchesofthehill/manabrew/compare/v3.21.0...v3.21.1) (2026-08-24)
+
+## [3.21.0](https://github.com/witchesofthehill/manabrew/compare/v3.20.0...v3.21.0) (2026-08-24)
+
+### Features
+
+* measure each leg of hosted latency separately ([#754](https://github.com/witchesofthehill/manabrew/issues/754)) ([31f1378](https://github.com/witchesofthehill/manabrew/commit/31f1378c899ed19654b3f7937408058a4450a868))
+
+## [3.20.0](https://github.com/witchesofthehill/manabrew/compare/v3.19.1...v3.20.0) (2026-08-24)
+
+### Features
+
+* add premodern format support ([#749](https://github.com/witchesofthehill/manabrew/issues/749)) ([91273d1](https://github.com/witchesofthehill/manabrew/commit/91273d16d68563d3a42e31852655ab71bf0916a8))
+
 ## [3.19.1](https://github.com/witchesofthehill/manabrew/compare/v3.19.0...v3.19.1) (2026-08-24)
 
 ### Fixes